demo-langchain/agent/requirements.txt
openkitx403-client==0.1.0
langchain==0.1.0
langchain-openai==0.0.5
solders==0.21.0
python-dotenv==1.0.0